requirements.txt
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
autopep8==1.5.6
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
constantly==15.1.0
cryptography==3.4.7
dateparser==1.0.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
numpy==1.20.2
pandas==1.2.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pyOpenSSL==20.0.1
python-binance==0.7.10
python-dateutil==2.8.1
pytz==2021.1
regex==2021.4.4
requests==2.25.1
service-identity==18.1.0
six==1.15.0
toml==0.10.2
Twisted==21.2.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
tzlocal==2.1
ujson==4.0.2
urllib3==1.26.4
Werkzeug==1.0.1
zope.interface==5.3.0
